requirements.txt
backcall==0.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-language==1.3.0
googleapis-common-protos==1.6.0
grpcio==1.24.0
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
jupyter-client==5.3.3
jupyter-core==4.5.0
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
protobuf==3.9.2
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.2.0
rope==0.14.0
rsa==4.0
six==1.12.0
tornado==6.0.3
traitlets==4.3.2
TwitterSearch==1.0.2
urllib3==1.25.6
wcwidth==0.1.7
